Refactor to move the model to a more accurate place
