+++ cloudoptimized/src/main/java/au/org/aodn/cloudoptimized/model/RelationType.java
@@ -1,6 +1,7 @@
-package au.org.aodn.esindexer.model;
+package au.org.aodn.cloudoptimized.model;
 
 import lombok.Getter;
+
 // TODO: Should use some lib provide value or move it to stacmodel folder
 @Getter
 public enum RelationType {
