+++ cloudoptimized/src/main/java/au/org/aodn/cloudoptimized/service/DataAccessService.java
@@ -1,11 +1,10 @@
-package au.org.aodn.esindexer.service;
-
-import au.org.aodn.esindexer.model.TemporalExtent;
-import au.org.aodn.stac.model.StacItemModel;
+package au.org.aodn.cloudoptimized.service;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import au.org.aodn.cloudoptimized.model.TemporalExtent;
+import au.org.aodn.stac.model.StacItemModel;
 
 public interface DataAccessService {
     List<StacItemModel> getIndexingDatasetBy(String uuid, LocalDate startDate, LocalDate endDate);
