--- cloudoptimized/src/main/resources/application.yaml
@@ -1,8 +1,0 @@
-cloudoptimized:
-  s3path: 's3a://imos-data-lab-optimised/parquet/raymond/'
-
-logging:
-  level:
-    root: INFO
-    org.apache.spark: WARN
-    org.apache.hadoop: WARN
