+++ indexer/pom.xml
@@ -222,18 +222,22 @@
         </dependency>
         <dependency>
             <groupId>au.org.aodn</groupId>
+            <artifactId>stacmodel</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>au.org.aodn</groupId>
+            <artifactId>cloudoptimized</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>au.org.aodn</groupId>
             <artifactId>ardcvocabs</artifactId>
             <version>${project.parent.version}</version>
             <classifier>tests</classifier>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>au.org.aodn</groupId>
-            <artifactId>stacmodel</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-
         <!-- https://mvnrepository.com/artifact/org.skyscreamer/jsonassert -->
         <dependency>
             <groupId>org.skyscreamer</groupId>
