+++ indexer/src/main/java/au/org/aodn/esindexer/service/GeoNetworkServiceImpl.java
@@ -1,9 +1,8 @@
 package au.org.aodn.esindexer.service;
 
+import au.org.aodn.cloudoptimized.model.RelationType;
 import au.org.aodn.esindexer.exception.MetadataNotFoundException;
-import au.org.aodn.esindexer.model.RelationType;
 import au.org.aodn.esindexer.utils.CommonUtils;
-import au.org.aodn.esindexer.utils.StringUtil;
 import au.org.aodn.esindexer.configuration.AppConstants;
 import au.org.aodn.esindexer.utils.UrlUtils;
 import au.org.aodn.stac.model.LinkModel;
