+++ indexer/src/main/java/au/org/aodn/esindexer/service/StacCollectionMapperService.java
@@ -1,7 +1,8 @@
 package au.org.aodn.esindexer.service;
 
+import au.org.aodn.cloudoptimized.model.RelationType;
+import au.org.aodn.cloudoptimized.service.DataAccessService;
 import au.org.aodn.esindexer.utils.AssociatedRecordsUtil;
-import au.org.aodn.esindexer.model.RelationType;
 import au.org.aodn.esindexer.utils.*;
 import au.org.aodn.stac.model.*;
