+++ pom.xml
@@ -17,9 +17,9 @@
     <description>Meta data indexer for Elastic Search</description>
 
     <modules>
-        <module>cloudoptimized</module>
         <module>geonetwork4-api</module>
         <module>stacmodel</module>
+        <module>cloudoptimized</module>
         <module>indexer</module>
         <module>ardcvocabs</module>
     </modules>
